Added Sustain Button
By tapping on the button in the bottom right corner of the controller,
you can then select a button that the emulator will "hold down" for
you. This helps in such circumstances as when you have to run and jump
in Mario; all you would have to do is tap the sustain button, tap the
run button, and then move in whatever direction you want and it will
automatically run for you.